+++ psl-core/src/main/java/org/linqs/psl/application/learning/weight/search/grid/InitialWeightGridSearch.java
@@ -64,7 +64,7 @@ public class InitialWeightGridSearch extends GridSearch {
     }
 
     @Override
-    protected double inspectLocation(float[] weights) {
+    protected double inspectLocation(double[] weights) {
         // Just have the internal WLA learn and then get the loss as the score.
         internalWLA.learn();
