+++ psl-core/src/main/java/org/linqs/psl/model/rule/logical/UnweightedLogicalRule.java
@@ -21,6 +21,7 @@ import java.util.List;
 
 import org.linqs.psl.model.atom.GroundAtom;
 import org.linqs.psl.model.formula.Formula;
+import org.linqs.psl.model.rule.Rule;
 import org.linqs.psl.model.rule.UnweightedRule;
 import org.linqs.psl.model.rule.WeightedRule;
 
@@ -34,7 +35,7 @@ public class UnweightedLogicalRule extends AbstractLogicalRule implements Unweig
     }
 
     @Override
-    public WeightedRule relax(float weight, boolean squared) {
+    public WeightedRule relax(double weight, boolean squared) {
         return new WeightedLogicalRule(formula, weight, squared, name);
     }
