+++ psl-core/src/main/java/org/linqs/psl/model/rule/logical/WeightedGroundLogicalRule.java
@@ -46,12 +46,12 @@ public class WeightedGroundLogicalRule extends AbstractGroundLogicalRule impleme
     }
 
     @Override
-    public float getWeight() {
+    public double getWeight() {
         return ((WeightedRule)rule).getWeight();
     }
 
     @Override
-    public void setWeight(float weight) {
+    public void setWeight(double weight) {
         ((WeightedRule)rule).setWeight(weight);
     }
 
@@ -95,7 +95,7 @@ public class WeightedGroundLogicalRule extends AbstractGroundLogicalRule impleme
             }
         }
 
-        WeightedLogicalRule newRule = new WeightedLogicalRule(rule.getFormula(), -1.0f * ((WeightedLogicalRule)rule).getWeight(), isSquared(), name);
+        WeightedLogicalRule newRule = new WeightedLogicalRule(rule.getFormula(), -1.0 * ((WeightedLogicalRule)rule).getWeight(), isSquared(), name);
         return new WeightedGroundLogicalRule(newRule, positiveAtoms, negativeAtoms, rvaCount);
     }
 }
